requirements.txt
spacy
pypdf
python-dotenv
pydantic
langchain
langchain-core
langchain-openai
langchain-google-genai
langchain-text-splitters
pytest
python-docx